line-deepl-bot/requirements.txt
Flask==3.0.3
line-bot-sdk==3.11.0
requests==2.32.3
python-dotenv==1.0.0
